refactor(exchange): extract GateStorage and decouple the gates from Action (AO-514)
KycGate and FeeGate were `is ExchangeStorage`, so the forthcoming primary
settlement contract could not use the compliance gate without also inheriting
the order book. They also spoke ExchangeTypes.Action, which is the exchange's
action set and not a second venue's.

- New GateStorage holds the four gate mappings (usedNonce, usedFeeNonce,
  complianceRequired, allowedCollectors) in ERC-7201 namespaced storage
  (assetera.storage.Gate), plus AccessControlUpgradeable/EIP712Upgradeable and
  the two errors the gates raise. ExchangeStorage is now `is GateStorage` and
  keeps only _orders/totalOrders/_offers/totalOffers/__gap.
- The four public mapping getters become explicit view functions over the
  namespaced struct. Same names, selectors and argument types.
- KycAttestation.action and FeeAttestation.action are retyped from Action to
  uint8, and moved with the structs into the new GateTypes. The digest and the
  calldata encoding are unchanged: the typehashes always said `uint8 action`
  and both encoders always cast. Proven by the digest vectors pinned in the
  previous commit, which pass unchanged.
- The inline "which actions may carry a paramsHash" enumeration in _verifyKyc
  becomes the virtual _paramsHashAllowed hook, defaulting to false, overridden
  in AsseteraECS with today's exact list.
- _verifyKyc keeps its complianceRequired early return; _verifyFee still has
  none (AC-884). Unchanged.

STORAGE LAYOUT IS INTENTIONALLY BROKEN. Moving the gate mappings out of the
linear layout shifts _offers 5->2, totalOffers 6->3 and __gap 8->4, so this
implementation is NOT upgrade-safe against the deployed Amoy/Sepolia proxies.
It is to be landed by a coordinated redeploy, never by upgradeToAndCall. The
snapshot is re-baselined accordingly.

Also adds GateStorageNamespace.t.sol, which re-derives the ERC-7201 constant
from its preimage and proves real writes land at the derived slots - a wrong
namespace constant is otherwise silent.

script/Verify.s.sol and the test suite carry type-signature adjustments only.

+++ contracts/script/Verify.s.sol
@@ -92,9 +92,9 @@ contract Verify is Script {
         );
 
         // 7. Compliance gating is on for all actions by default.
-        bool placeGated = exchange.complianceRequired(ExchangeTypes.Action.Place);
-        bool fillGated = exchange.complianceRequired(ExchangeTypes.Action.Fill);
-        bool settleGated = exchange.complianceRequired(ExchangeTypes.Action.Settle);
+        bool placeGated = exchange.complianceRequired(uint8(ExchangeTypes.Action.Place));
+        bool fillGated = exchange.complianceRequired(uint8(ExchangeTypes.Action.Fill));
+        bool settleGated = exchange.complianceRequired(uint8(ExchangeTypes.Action.Settle));
         _check(
             "all actions KYC-gated by default",
             placeGated && fillGated && settleGated,
